+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - feat: add predict algo category (#201)
 
+### Changes
+
+- BREAKING CHANGE (UpdateComputePlanSpec key field): Use new register tasks endpoint (#196)
+
 ## [0.23.0](<https://github.com/owkin/substra/releases/tag/0.23.0>) - 2022-05-31
 
 ### Added